requirements.txt
django
djangorestframework
django-rest-knox
django-polymorphic